chore: moved gpt_chat_completion to a dedicated module
