+++ aidial_adapter_openai/utils/exceptions.py
@@ -29,7 +29,7 @@ class HTTPException(Exception):
 
 def create_error(
     message: str, type: str | None, param: Any = None, code: Any = None
-):
+) -> dict:
     return {
         "error": {
             "message": message,
